Migrate MTP backend to the backend-neutral mtp-rs API
Repoints Cmdr onto mtp-rs's new backend-neutral `mtp::` surface so the MTP
integration no longer speaks PTP-over-USB vocabulary. This decouples Cmdr from
the transport and is the prerequisite for mtp-rs's upcoming Windows WPD backend.

- Error mapping: `connection/errors.rs` rewritten against the neutral, non-exhaustive
  `mtp_rs::Error` (`NotFound`/`StaleHandle`/`AccessDenied`/`ExclusiveAccess`/`Busy`/
  `StorageFull`/`Io`/`Other`/...). Stale-handle detection via `e.is_stale_handle()`,
  exclusive access via `e.is_exclusive_access()`; the raw `nusb` `ErrorKind` is gone.
- Neutral types: directory checks use `ObjectInfo::is_folder()` (not the PTP
  `ObjectFormatCode::Association`); storage reads use `is_writable` / `total_capacity`
  / `free_space`; write capability comes from `device.capabilities().can_upload`
  (not `supports_operation`); datetimes/events use the neutral `DateTime` / `DeviceEvent`.
- Opaque ids: `ObjectHandle`/`StorageId` are now `u64`. Cmdr keeps its u32 wire DTOs,
  narrowing at the boundary (`StorageId(u64::from(id))`, `handle.0 as u32`).
- Downloads: offset reads go through `download_windowed(h, ByteRange::From(off), ws)`.
- Upload progress: mtp-rs's `Storage::upload` requires a `'static` data stream, but
  Cmdr's `write_from_stream` carries a borrowed progress callback. Keep the callback
  out of the stream — a `'static` chunk stream reports byte totals over a channel and
  watches a shared cancel flag, while a relay loop owns the borrow and drives
  `on_progress`. (See the note to the lead: relaxing mtp-rs's `'static` bound would
  let this revert to the simpler borrowed form.)
- Test fix: the virtual device now surfaces its serial through discovery, so its
  canonical id is serial-based; the scan-oracle test now derives the device id from
  discovery instead of hardcoding `mtp-{location_id}`.

Dep is a TEMPORARY path pin to the local mtp-rs validation worktree; revert to a
crates.io version once the neutral API ships.

Default suite green (3029 passed); virtual-mtp MTP tests pass run serially.

+++ apps/desktop/src-tauri/src/indexing/mtp_watch.rs
@@ -390,7 +390,7 @@ fn resolve_and_apply_buffered_handles(volume_id: &str, handles: Vec<(u32, u32)>)
     tauri::async_runtime::spawn(async move {
         for (storage_id, handle) in handles {
             match crate::mtp::connection::connection_manager()
-                .resolve_object_for_index(&device_id, storage_id, mtp_rs::ObjectHandle(handle))
+                .resolve_object_for_index(&device_id, storage_id, mtp_rs::ObjectHandle(u64::from(handle)))
                 .await
             {
                 Ok(obj) => apply_mtp_added_or_changed(
